requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.26.2
psycopg2-binary==2.9.7
gunicorn==21.2.0
pillow==10.0.1